1.1.4.beta 250422 **✨ Update** - ✨ Progressing on the product module; implementation is not yet complete and has not been tested. - ✨ 正在推进商品模块，尚未编写完毕且尚未测试。

+++ kitex server/items/model/item.go
@@ -3,10 +3,10 @@ package model
 import "time"
 
 type Item struct {
-	Id        int
+	Id        int64
 	Name      string
 	Price     float64
-	Stock     int
+	Stock     int64
 	Intro     string
 	CreatedAt time.Time
 	UpdatedAt time.Time
